pyproject.toml
[project]
name = "nala-neat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "gymnasium>=1.2.3",
    "hydra-core>=1.3.2",
    "ray>=2.53.0",
    "torch>=2.10.0",
]
